requirements.txt
attrs==18.2.0
Click==7.0
python-dotenv==0.10.0
